requirements.txt
Flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
openai==1.6.1
python-dotenv==1.0.0
reportlab==4.0.8
